MAINT: Move fracture meshing helper classes to fracture network module
This will allow for their adaptation also for 2d fracture meshing
